#540 lineage/impact overview JS generalization
